Added all blog-related features to the app

+++ next-app/src/app/api/admin/blog/route.ts
@@ -1,5 +1,6 @@
 import { auth } from "@/auth";
 import prisma from "@/config/db";
+import { revalidatePath } from "next/cache";
 import { NextRequest, NextResponse } from "next/server";
 
 export const POST = async (req: NextRequest) => {
@@ -43,27 +44,55 @@ export const POST = async (req: NextRequest) => {
 
 export const PUT = async (req: NextRequest) => {
     const body = await req.json();
+    const session = await auth();
     try {
-        const blog = await prisma.blog.update({
+        const blog = await prisma.blog.findUnique({
+            where: {
+                id: body.id
+            }
+        });
+
+        if (!blog) {
+            return NextResponse.json({ error: "Blog not found." }, { status: 404 });
+        }
+        if (session?.user.id !== blog?.authorId) {
+            return NextResponse.json({ error: "You are not authorized to update this blog." }, { status: 403 });
+        }
+
+        await prisma.blog.update({
             where: {
                 id: body.id
             },
             data: body
         })
-        return NextResponse.json(blog, { status: 200 });
+        revalidatePath(`/blog/${blog.slug}`);
+        return NextResponse.json({ message: "Blog updated." }, { status: 200 });
     } catch (error: any) {
         return NextResponse.json({ error: error.message }, { status: 500 });
     }
 }
 
 export const DELETE = async (req: NextRequest) => {
+    const body = await req.json();
     try {
-        const body = await req.json();
+        const session = await auth();
+        const blog = await prisma.blog.findUnique({
+            where: {
+                id: body.id
+            }
+        });
+
+        if (session?.user.id !== blog?.authorId) {
+            return NextResponse.json({ error: "You are not authorized to delete this blog." }, { status: 403 });
+        }
+
         await prisma.blog.delete({
             where: {
                 id: body.id
             }
         });
+
+        revalidatePath(`/blog/${blog?.slug}`);
         return NextResponse.json({ message: "Blog deleted successfully" }, { status: 200 });
     } catch (error: any) {
         return NextResponse.json({ error: error.message }, { status: 500 });
